// apps/portal/src/api/generated/model/assemblyRecordPreviewDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */
import type { AssemblyRecordPreviewDtoMajorityComparator } from "./assemblyRecordPreviewDtoMajorityComparator";
import type { AssemblyRecordPreviewDtoOutcome } from "./assemblyRecordPreviewDtoOutcome";
import type { FractionDto } from "./fractionDto";
import type { RulesetFractionDto } from "./rulesetFractionDto";

export interface AssemblyRecordPreviewDto {
    outcome: AssemblyRecordPreviewDtoOutcome;
    majorityMet: boolean;
    /** @nullable */
    winningOptionId: string | null;
    majorityThreshold: RulesetFractionDto;
    majorityComparator: AssemblyRecordPreviewDtoMajorityComparator;
    majorityDenominator: FractionDto;
}